% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_measures_lte.R
\name{all_measures_lte}
\alias{all_measures_lte}
\title{Test if all measures are nless-than-or-equal to previous measure.}
\usage{
all_measures_lte(measures_before, measures_now)
}
\arguments{
\item{measures_before}{numeric; measures between the second most recently
completed data set and its predecessor.}

\item{measures_now}{numeric; measures between the most recently completed
dataset and its predecessor.}
}
\value{
logical; \code{TRUE} if all new values are not less than the previous
values.
}
\description{
The stop condition for the main algorithm loop in missForest (Stekhoven and
Buehlmann, 2012) requires that all measures have either decreased or
increased. This is a helper function that tests if all measures are less than
or equal to the previous value.
}
\details{
Given two (vectors of) measures of the 'change' between iterations for the
past two iterations of the missForest loop, this should return \code{TRUE} when
the loop should break. Another function, for example
\code{\link[=measure_stekhoven_2012]{measure_stekhoven_2012()}}, calculates the measures passed to this test. The
missForest measures are, for example;
\enumerate{
\item the \emph{negative} sum of square differences between the imputed values (of
continuous variables), and;
\item the proportion of stationary values (for categorical data).
}

Alternatively, we could use;
\enumerate{
\item the mean of the (rank) correlation between the datasets of each ordered
variable, and;
\item the proportion of stationary values (for non-ordered, categorical data).
}

This treats all ordered data the same way, and should identify when the
entropy is close to a local optimum, see \code{\link[=measure_correlation]{measure_correlation()}}.
}
\references{
\itemize{
\item Stekhoven, D. J. & Buehlmann, P. (2012). MissForest--non-parametric
missing value imputation for mixed-type data. \emph{Bioinformatics}, \emph{28}(1),
112-118. \url{doi:10.1093/bioinformatics/btr597}
}
}
\seealso{
\code{\link[=measure_correlation]{measure_correlation()}} \code{\link[=measure_stekhoven_2012]{measure_stekhoven_2012()}}
}
