% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mids.R
\name{as_mids}
\alias{as_mids}
\title{Convert imputed datasets to \code{mids} object.}
\usage{
as_mids(x)
}
\arguments{
\item{x}{list (named, class 'mimputest'): imputed datasets as returned by
\code{\link[=mimputest]{mimputest()}}.}
}
\value{
a mids object, see \link[mice:mids-class]{mice::mids} for details.
}
\description{
To take advantage of generics and other functions available in 'mice' (and
other packages) the output from \code{\link[=mimputest]{mimputest()}} can be converted into a \code{mids}
object. Some functionality may be broken due to non-equivalence of
'mimputest' and 'mice' syntax.
}
\details{
See \code{\link[mice:mids-class]{mice::mids()}} for details on the \code{mids} class. This function uses
several 'mice' functions to construct a skeleton \code{mids} object that reflects
a near-equivalent call in 'mice' to generate the imputed data. Some
functionality for \code{mids} objects may be missing because an exact equivalent
is not possible. For example \code{blots} and \code{post} are provided but not using
syntax that \code{\link[mice:mice]{mice::mice()}} would accept. \code{seed} is recorded as per
\code{\link[=mimputest]{mimputest()}} and \code{lastSeedValue} is set to \code{NULL}. The call stored is the
original call to \code{\link[=mimputest]{mimputest()}}.
}
\seealso{
\link[mice:mids-class]{mice::mids}
}
