% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_results.R
\name{complete_results}
\alias{complete_results}
\title{Completed data sets as long data.frame}
\usage{
complete_results(
  X,
  mf,
  k = length(mf$results),
  to.complete = names(mf$which_imputed),
  replace = F
)
}
\arguments{
\item{X}{data.frame;
incomplete data set passed to \code{\link{smirf}} in construction of
\code{mf}.}

\item{mf}{list;
imputed data sets and output of missForest procedure as returned by
\code{\link{smirf}}.}

\item{k}{integer;
number of completed data sets to generate, default is the number of
imputed data sets in \code{mf}.}

\item{to.complete}{character;
vector of names of variables to include, default is all variables in
imputed data sets in \code{mf}.}

\item{replace}{boolean;
indicator of sampling from imputed data sets with replacement or not.}
}
\value{
data.frame;
    completed long-format data containing the columns of \code{X} and two
    more columns, \describe{
       \item{\code{.source}}{integer; id of the source data set in \code{mf}
           of the imputed data in the row (if any), otherwise \code{NA},
           and;}
       \item{\code{.imp}}{integer; id of the complete data set between 1 and
           \code{k}, or \code{NA} if it belongs to all (i.e. contains no
           imputed data),}
    } and each row is either \itemize{
        \item a complete case whose values are unaltered, or;
        \item an incomplete case with a sample from the imputed data
              replacing any missing values.
    }
}
\description{
Completed (imputed) data sets given incomplete data passed to
\code{\link{smirf}} and the output thereof, returned in a long-format
data.frame.
}
\details{
Description
}
\section{To-do}{

\itemize{
  \item document \code{complete_results}
  \item write tests for \code{complete_results}
}
}

\seealso{
\code{\link{smirf}}
}
