% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Calculate entropy of categorical data}
\usage{
entropy(x, freq = table(x))
}
\arguments{
\item{x}{ordered;
observed categorical data.}

\item{freq}{numeric;
count of observations in each group.}
}
\value{
numeric;
            scalar value of entropy.
}
\description{
Calculate entropy of categorical data using tally of observations in each
category.
}
\section{To-do}{

\itemize{
  \item document \code{entropy}.
  \item write tests for \code{entropy}.
}
}

\keyword{internal}
