% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maps_to_categories.R
\name{get_maps_to_categories}
\alias{get_maps_to_categories}
\title{Map from levels proposed by as.factor to each integer and logical column}
\usage{
get_maps_to_categories(X)
}
\arguments{
\item{X}{data.frame;
a incomplete data set including any of numeric, logical, integer,
factor and ordered data types.}
}
\value{
named list;
           names are given by the column names of \code{X} which have logical
           or integer data, and each item is either a named logical or named
           integer vector, with names given by the unique factor levels
           due to passing the column to \code{\link{as.factor}} and the
           values are given by the corresponding unique values of the column,
}
\description{
Creates maps from the levels proposed by \code{\link{as.factor}} when it is
applied to each integer and logical column in a data.frame. The maps are
given in a list with names given by the names of the integer and logical
columns in the data set. Each item is either a named logical or named integer
vector, with names given by the unique factor levels, and values given by the
unique values of the corresponding column.
}
\seealso{
\code{\link{as.factor}}
}
\keyword{internal}
