% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_naive.R
\name{impute_naive}
\alias{impute_naive}
\title{Impute missing data naively using complete cases.}
\usage{
impute_naive(
  data,
  indicator = lapply(data, is.na),
  mode = c("sample", "aggregate")
)
}
\arguments{
\item{data}{data.frame: an incomplete data set of numeric, logical, integer,
factor, or ordered data - one row per observation, one column per variable.}

\item{indicator}{list: an indicator of missing status for each (named)
variable for each row in the data, equal to 1 (or \code{TRUE}) for a missing case,
otherwise 0 (or \code{FALSE}).}

\item{mode}{character: whether to draw randomly from observed cases
(\verb{='sample'}) or to use aggregates such as the mean or most-frequent-value
(\verb{='aggregate'}).}
}
\value{
data.frame: an imputed data set with the same structure as the \code{data}
argument.
}
\description{
Drawing imputed values without using the fully conditional specification (or
other joint model) is performed during the first step of both the 'mice'
imputation algorithm (van Buuren and Groothuis-Oudshoorn, 2012) and the
'missForest' estimation algorithm (Stekhoven and Buehlmann, 2012). For the
former, sampling from complete cases is used, whereas for the latter, the
most-frequent value or mean of the complete cases is used.
}
\details{
Two wrappers are provided for this function so that it can be passed as
\code{fn_init} to \code{\link[=mimputest]{mimputest()}}, these are \code{\link[=impute_naive_by_sample]{impute_naive_by_sample()}} and
\code{\link[=impute_naive_by_aggregate]{impute_naive_by_aggregate()}}. In either case, the function will modify each
column (in order) by replacing missing values with either a sample from the
complete cases, or an aggregate of the complete cases (such as the most
frequent value). The former is for 'mice'-like imputation, the latter is for
'missForest'.

Sampling generates data with higher entropy than given by the aggregate
approach. When using a stopping condition, care must be taken that the choice
of initial state doesn't introduce a false positive 'stop' in early
iterations. We suspect that the 'aggregate' approach will results in better
specificity of the stop condition.
}
\examples{
impute_naive(data.frame(x=c(0,1,NA)), mode='sample')
impute_naive(data.frame(x=c(0,1,NA)), mode='aggregate')

}
\references{
\itemize{
\item Stekhoven, D.J. & Buehlmann, P. (2012). MissForest--non-parametric
missing value imputation for mixed-type data. \emph{Bioinformatics}, \emph{28}(1),
112-118. \doi{10.1093/bioinformatics/btr597}.
\item Van Buuren, S. & Groothuis-Oudshoorn, K. (2011). mice: Multivariate
Imputation by Chained Equations in R. \emph{Journal of Statistical Software},
\emph{45}(3), 1-67. \doi{10.18637/jss.v045.i03}.
}
}
\seealso{
\code{\link[=impute_naive_by_sample]{impute_naive_by_sample()}} \code{\link[=impute_naive_by_aggregate]{impute_naive_by_aggregate()}}
}
