% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_naive.R
\name{impute_naive_by_aggregate}
\alias{impute_naive_by_aggregate}
\title{Impute missing data naively using aggregate of complete cases.}
\usage{
impute_naive_by_aggregate(data, indicator = lapply(data, is.na))
}
\arguments{
\item{data}{data.frame: an incomplete data set of numeric, logical, integer,
factor, or ordered data - one row per observation, one column per variable.}

\item{indicator}{list: an indicator of missing status for each (named)
variable for each row in the data, equal to 1 (or \code{TRUE}) for a missing case,
otherwise 0 (or \code{FALSE}).}
}
\value{
data.frame; an imputed data set with the same structure as the \code{data}
argument.
}
\description{
Drawing imputed values without using the fully conditional specification (or
other joint model) is performed during the first step of the 'missForest'
estimation algorithm (Stekhoven and Buehlmann, 2012). In this case, the
the most-frequent value or mean of the complete cases is used.
}
\details{
This function can be passed directly to \code{\link[=mimputest]{mimputest()}} as the value of the
\code{fn_init} argument. This function will iterate over each column in the data
and replace missing values in one of two ways depending on the type of the
data;
\itemize{
\item if the data are continuous then the mean of the complete cases is used;
\item if the data are categorical (including logical, integers, etc) then the
most frequent value is used, in the case of ties a single values is
randomly selected from the set of most frequent values.
}

This approach has lowest possible entropy, this may improve the specificity
(but reduce the sensitivity) of some stopping conditions in the main
algorithm loop during the earlier iterations.
}
\examples{
\dontrun{
mimputest(iris, fn_init=impute_naive_by_aggregate)
}
impute_naive_by_aggregate(data.frame(x=c(0,1,NA)))

}
\references{
\itemize{
\item Stekhoven, D.J. & Buehlmann, P. (2012). MissForest--non-parametric
missing value imputation for mixed-type data. \emph{Bioinformatics}, \emph{28}(1),
112-118. \doi{10.1093/bioinformatics/btr597}.
}
}
\seealso{
\code{\link[=impute_naive]{impute_naive()}}
}
