% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_naive.R
\name{impute_naive_by_sample}
\alias{impute_naive_by_sample}
\title{Impute missing data naively by sampling from complete cases}
\usage{
impute_naive_by_sample(data, indicator = lapply(data, is.na))
}
\arguments{
\item{indicator}{list (named); each item is the indicator (vector) of missing
values for a column in the data.}
}
\value{
data.frame; an imputed data set with the same structure as the \code{data}
argument.
}
\description{
The initial step in multiple imputation with random forests requires missing
values to be drawn \emph{without} using the model; in the case of multiple
imputation via chained equations the complete cases are sampled from with
replacement (van Buuren and Groothuis-Oudshoorn, 2012).
}
\details{
#' This function can be passed directly to \code{\link[=smirf]{smirf()}} as the value of the
\code{fn_init} argument. This function will iterate over each column in the data
and replace missing values by sampling (with replacement) from the complete
cases.

This approach may improve the sensitivity (but lower the specificity) in some
stopping conditions used by the main algorithm loop during early iterations,
owing to the entropy being closer to that of the converged process.
}
\examples{
\dontrun{
smirf(iris, fn_init=impute_naive_by_aggregate)
}
impute_naive_by_sample(data.frame(x=c(0,1,NA)))

}
\references{
\itemize{
\item Van Buuren, S. & Groothuis-Oudshoorn, K. (2011). mice: Multivariate
Imputation by Chained Equations in R. \emph{Journal of Statistical Software,
45}(3), 1-67. \url{doi:10.18637/jss.v045.i03}.
}
}
\seealso{
\code{\link[=impute_naive]{impute_naive()}} \code{\link[mice:mice]{mice::mice()}}
}
