% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_correlation.R
\name{measure_correlation}
\alias{measure_correlation}
\title{Measure correlation and stationary proportion between completed data sets}
\usage{
measure_correlation(X, Y, X_init, indicator, method = "spearman")
}
\arguments{
\item{X}{named list;
imputed values, in order of appearance by row in original data,
of each variable (named) from one iteration within missForest
procedure.}

\item{Y}{named list;
imputed values, in order of appearance by row in origina data,
of each variable (named) from the iteration within the missForest
procedure succeeding that used to determine \code{X}.}

\item{X_init}{data.frame;
a data set including any of numeric, logical, integer, factor and
ordered data types, to be used as the initial state of the
missForest procedure.}

\item{indicator}{named list;
an indicator of the missing (\code{=T}) or not-missing (\code{=F})
status of the columns of \code{X_init}.}

\item{method}{character;
passed to \code{\link[stats]{cor}}.}
}
\value{
named numeric;
            two named values: \describe{
                \item{\code{continuous}}{mean (rank) correlation of the
                     continuous and ordered data between the two completed
                     data sets, and;}
                \item{\code{categorical}}{proportion of stationary values of
                    categorical (unordered) data between the two completed
                    data sets (see \code{\link{stationary_rate}}).}
            }
}
\description{
Measures a relationship between two supplied completed data sets, typically
generated by two sequential iterations of the missForest procedure. 
Given by the rank correlation or proportion of stationary values for
ordered (including continuous) and categorical data respectively.
}
\details{
Measures a relationship between two supplied completed data sets, typically
generated by two sequential iterations of the missForest procedure. Intended
to be used with the stop criterion that as soon as all values (see below)
remain constant or decrease at once, then the missForest procedure is deemed
to have converged.

The two values are the;
\itemize{
   \item mean rank correlation of the ordered (including continuous) data
         between the supplied completed data sets, and;
   \item proportion of stationary categorical data between the two supplied
         completed data sets (see \code{\link{stationary_rate}}).
}

The type of correlation can be modified to that of \code{kendall} or
\code{pearson} (not recommended for ordered data) by the \code{method}
argument passed on to \code{\link[stats]{cor}}.
}
\examples{
\dontrun{
# simply pass to miForang
miForang(iris, stop.measure=measure_correlation)
}
}
\seealso{
\code{\link[stats]{cor}} \code{\link{miForang}}
         \code{\link{stationary_rate}}
}
