% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure_stekhoven_2012}
\alias{measure_stekhoven_2012}
\title{Measure differences between sequential complete data sets according to
missForest stopping condition.}
\usage{
measure_stekhoven_2012(x_sample, y_sample, data, indicator)
}
\arguments{
\item{x_sample}{named list: the imputed values of the missing cases within
each variable from one iteration of the missForest algorithm.}

\item{y_sample}{named list: the imputed values from the following (or
preceding) iteration for \code{x_sample} with the same format as \code{x_sample}.}

\item{data}{data.frame: an incomplete data set of numeric, logical, integer,
factor, or ordered data - one row per observation, one column per variable.}

\item{indicator}{list: an indicator of missing status for each (named)
variable for each row in the data, equal to 1 (or \code{TRUE}) for a missing case,
otherwise 0 (or \code{FALSE}).}
}
\value{
named numeric; two name values \describe{
\item{\code{continuous}}{a Frobenius norm-based relative difference of the
continuous data between the two complete data sets, and;}
\item{\code{categorical}}{the proportion of stationary values of categorical,
including ordered, data between the two complete data sets.}
}
}
\description{
The differences between one complete data set to another are used to assess
the stopping criterion for the missForest procedure (Stekhoven and
Buehlmann, 2012). Up to two measures are calculated, the first is a Frobenius
norm-based measure of the distance between continuous data, the second is the
proportion of stationary values for categorical data.
}
\details{
The missForest algorithm compares sequential sets of measures, i.e. the
measured difference between the most recent and second-most recent data sets
versus that between the second and third-most recent data sets. When all
measures have changed direction, the algorithm stops. It is assumed that the
measures are initially increasing, and when the measure decreases the
algorithm has converged.

The measures per data set, specifically, are:
\itemize{
\item the negative of the sum of the differences squares of all continuous
variables divided by the sum of the squares of the continuous data in the
latest data set, and;
\item the proportion of stationary values  of categorical data divided by the
total number of missing and non-missing values (of categorical data).
}

These differ slightly from the formulas in the original missForest paper, as
we use a decrease between successive measures as the indicator for stopping,
and the categorical measure stated in the paper contains a typographical
error, see the source code \url{https://github.com/stekhoven/missForest/}
(commit \code{eb2f018}) for correct formula.
}
\examples{
\dontrun{
# simply pass to mimputest
mimputest(iris, stop_measure=measure_stekhoven_2012)
}

}
\references{
\itemize{
\item Stekhoven, D. J. & Buehlmann, P. (2012). MissForest--non-parametric
missing value imputation for mixed-type data. \emph{Bioinformatics}, 28(1),
112-118. \doi{10.1093/bioinformatics/btr597}.
}
}
\seealso{
\code{\link[=sampler_loop]{sampler_loop()}}
}
