% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_ranger.R
\name{sample_from_ranger}
\alias{sample_from_ranger}
\title{Predict or sample values from fitted ranger object}
\usage{
sample_from_ranger(ranger_fit, data_, response, tree.imp = F, ...)
}
\arguments{
\item{ranger_fit}{ranger object;
random forest fitted to data of same format as \code{data_}
argument.}

\item{data_}{a data.frame;
data.frame with same variables as the data used in the training
of \code{ranger_fit}.}

\item{response}{character;
name of the dependent variable i.e. the value of
\code{dependent.variable.name} from the corresponding call to
\code{\link{ranger}} in the fitting step.}

\item{tree.imp}{logical;
set to \code{T} to impute using a single (randomly selected with
replacement) tree for each row in \code{data_} or \code{F}
(default) to use the aggregated 'whole-of-forest' prediction.}

\item{...}{additional arguments passed on to \code{predict.ranger}, not
including the \code{predict.all} or \code{num.trees} argument.}
}
\value{
predicted or sampled values or classes (with original class and
        levels)
}
\description{
Determines the predicted value using the 'whole-of-forest' prediction, or
evaluates the prediction of a selected tree (\code{tree.imp=T}). In the
latter case, each row in the data set is predicted using a randomly selected
(with replacement) tree in the forest. We call this a 'tree-sampled'
prediction and it \emph{somewhat} corresponds to an sample from the
distribution of the predicted value or class given the (new) data.
}
\details{
Ranger is considered slow at prediction as of 07/2018, particularly for
smaller data sets, see issue 133:
\url{https://github.com/imbs-hl/ranger/issues/133}
}
\seealso{
\code{\link[ranger]{ranger}}
}
\keyword{internal}
