% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_step.R
\name{sampler_step}
\alias{sampler_step}
\title{Perform one iteration of the procedure to generate an imputed data set.}
\usage{
sampler_step(
  data,
  model,
  indicator,
  call_lr_train,
  sampler,
  prediction_type,
  overrides,
  clean_step
)
}
\arguments{
\item{model}{matrix;
logical matrix which indicates inclusion of a predictor (named
column) in the model of an imputed value (named row), with the
order of imputation being the row order, default is a matrix of
ones with rows for each partially but not-completely missing
variable (in order of least to most missing), and columns
for every partially complete variable.}

\item{call_lr_train}{call: a 'skeleton' call to \code{\link[literanger:train]{literanger::train()}} for
fitting random forests, arguments can be over-ridden on a per-variable basis
by \code{overrides}.}

\item{overrides}{named list;
(variable-wise) over-rides for arguments passed to
\code{\link[ranger]{ranger}} when training on the response
variable given by the name of the item.}
}
\value{
named list: \describe{
\item{\code{oob_error}}{data.frame: out-of-bag error estimates for each
modelled variable with columns \code{iteration} (the identified of the
iteration), \code{variable} (name), \code{measure} (the type of OOB-error measure;
'mse' is mean square error, 'pfc' is proportion falsely classified), and
\code{value} is the estimate of the out-of-bag error.}
\item{\code{imputed}}{named list: the imputed values of missing cases for each
variable in \code{data}.}
}
}
\description{
Each step in the iterative procedure to generate an imputed data set results
in a new data set by training and predicting (using random forests) the
missing values for each variable.
}
\details{
A description of the overall process is given in \code{\link[=sampler_loop]{sampler_loop()}}. This is a
function internal to 'smirf' that calls the training and prediction methods
from 'literanger'. The arguments are described in greater detail in
\code{\link[=sampler_loop]{sampler_loop()}}. At the end of the step the out-of-bag error and the
imputed values (per variable) are returned.
}
\seealso{
\code{\link[=sampler_loop]{sampler_loop()}}
}
