% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_as_matrix.R
\name{samples_as_matrix}
\alias{samples_as_matrix}
\title{Random samples from the first natural numbers}
\usage{
samples_as_matrix(n, k = n, size = 0, replace = F)
}
\arguments{
\item{n}{integer;
defines set to sample from as \eqn{\{1, \ldots, n\}}.}

\item{k}{integer;
number of items to draw from the set \eqn{\{1, \ldots, n\}} in each
sample (set), also the number of columns in output.}

\item{size}{integer;
number of (sample) sets in total, also the number of rows in output.}

\item{replace}{boolean;
indicate whether to sample with replacement (within the set).}
}
\value{
matrix;
    a \code{size} by \code{k} matrix with a random sample from
    \eqn{\{1, \ldots, n\}} in each row.
}
\description{
Generate a matrix in which each row is a random sample from the set of
the first \code{n} natural numbers.
}
\details{
This is the author's best effort at efficiently generating a large number of
samples of fixed length (\code{k}) from the set of the first \code{n} natural
numbers without replacement. This implementation samples from the set of all
combinations when \code{n}, \code{k} and \code{size} are favourable,
otherwise it uses a straightforward call to \code{\link{base}{sapply}} and
\code{\link{base}{sample.int}}.

When sampling with replacement, the implementation is a trivial call to
\code{link{sample.int}} and a matrix constructor.
}
\seealso{
\code{\link{utils}{combn}} \code{\link{base}{sample.int}}
}
\keyword{internal}
