% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{stationary_rate}
\alias{stationary_rate}
\title{Compute rate of stationary values}
\usage{
stationary_rate(X, Y, X_init, indicator)
}
\arguments{
\item{X}{named list;
imputed values, in order of appearance by row in original data,
of each variable (named) from one iteration within missForest
procedure.}

\item{Y}{named list;
imputed values, in order of appearance by row in origina data,
of each variable (named) from the iteration within the missForest
procedure succeeding that used to determine \code{X}.}

\item{X_init}{data.frame;
a data set including any of numeric, logical, integer, factor and
ordered data types, to be used as the initial state of the
missForest procedure.}

\item{indicator}{named list;
an indicator of the missing (\code{=T}) or not-missing (\code{=F})
status of the columns of \code{X_init}.}
}
\value{
numeric; the total proportion of stationary values across the two
completed data sets.
}
\description{
Computes the rate of stationary values from one completed data set to another
where the non-missing data is included in the denominator of the calculation,
and - by definition - does not contribute to the numerator.
}
