% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{statistics_of_imputed}
\alias{statistics_of_imputed}
\title{Calculate statistics of imputed data}
\usage{
statistics_of_imputed(imputed)
}
\arguments{
\item{imputed}{named list; each item in the last contains the imputed values
of the named item.}
}
\value{
data.frame; a data.frame with three columns;
\describe{
\item{\code{variable}}{name of the imputed data as a factor;}
\item{\code{value}}{value of the measure/statistic;}
\item{\code{measure}}{the name of the measure as a factor, e.g. 'entropy'.}
}
}
\description{
Calculates selected statistics (measures) of imputed data
}
\details{
Calculates statistics or measures for each imputed variable according to its
type, which is one of categorical, ordinal or continuous:
\describe{
\item{categorical}{
calculate the \code{\link{entropy}} of non-numeric or integer data;
}
\item{ordinal}{
calculate the dispersion using \code{\link{leiks_D}} of ordered or
integer data;
}
\item{continuous}{
calculate the mean and variance of numeric data;
}
}

These statistics are intended to be used to monitor convergence of the
missForest or MICE procedure.
}
\seealso{
entropy leiks_D
}
\keyword{internal}
