% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_condition.R
\name{stop_condition}
\alias{stop_condition}
\title{Evaluate stop condition of imputation procedure}
\usage{
stop_condition(previous_measure, measure)
}
\arguments{
\item{previous_measure}{numeric;
values of a measure of the relationship between the second most
recent completed data set and that of iteration prior to it.}

\item{measure}{numeric;
values of a measure of the relationship between the most recent
completed data set and that of the iteration prior to it.}
}
\value{
logical;
            \code{T} if all values of \code{previous_measure} are greater
            than or equal to corresponding value of \code{measure}.
}
\description{
Trivial evaluation of two measures of the inter-iteration relationship
between completed data sets. Evaluates to \code{T} if all of the measurements
at the current iteration are comparable to and (strictly) less than the
measurements at the previous iteration, otherwise returns \code{F}.
}
\seealso{
\code{\link{measure_correlation}}
         \code{\link{measure_stekhoven_2012}}
}
