% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmap_categories.R
\name{unmap_categories}
\alias{unmap_categories}
\title{Map categorical data to values}
\usage{
unmap_categories(X, to_categories)
}
\arguments{
\item{X}{named list or data.frame;
list or data.frame of factor data}

\item{to_categories}{named list;
each named item is of the value-type to map the item or column
with the same name \code{X} to, with the names of each value being
the label of the (corresponding) factor to map from, e.g. to map
the factor levels 'A' and 'B' to integers 1 and 2 respectively 
for a column called \code{dummy}, then
\code{to_categories=list(dummy=c(A=1L,B=2L))}}
}
\value{
list or data.frame;
           list or data.frame of the same shape as \code{X} with factors
           mapped to the values specified in \code{to_categories}.
}
\description{
Map categorical data to specified values of new (or same) type.
}
\keyword{internal}
