% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_results.R
\name{complete_results}
\alias{complete_results}
\title{Generate rows of a matrix sampled from the first `n` natural numbers.}
\usage{
complete_results(
  X,
  mf,
  k = length(mf$results),
  to.complete = names(mf$which_imputed),
  replace = F
)
}
\arguments{
\item{X}{data.frame;
incomplete data set passed to \code{\link{smirf}} in construction of
\code{mf}.}

\item{mf}{list;
imputed data sets and output of missForest procedure as returned by
\code{\link{smirf}}.}

\item{k}{integer;
number of completed data sets to generate, default is the number of
imputed data sets in \code{mf}.}

\item{to.complete}{character;
vector of names of variables to include, default is all variables in
imputed data sets in \code{mf}.}

\item{replace}{boolean;
indicator of sampling from imputed data sets with replacement or not.}
}
\value{
data.frame;
    completed long-format data containing the columns of \code{X} and two
    more columns, \describe{
       \item{\code{.source}}{integer; id of the source data set in \code{mf}
           of the imputed data in the row (if any), otherwise \code{NA},
           and;}
       \item{\code{.imp}}{integer; id of the complete data set between 1 and
           \code{k}, or \code{NA} if it belongs to all (i.e. contains no
           imputed data),}
    } and each row is either \itemize{
        \item a complete case whose values are unaltered, or;
        \item an incomplete case with a sample from the imputed data
              replacing any missing values.
    }
}
\description{
Imputed datasets are generated once the chained equations loop has finished.
Each observation that has one or more imputed value is assigned an index
which determines which chain any imputed value is taken from. This function
generates a matrix of these indices, with each row corresponding to an
observation and each column corresponding to one imputed data set.
}
\details{
Within each imputed data set, each observation with a missing value is
assigned one chain to take values from. Typically the number of imputed data
sets is equal to the number of chains, and the chain indices are drawn
without replacement for each observation. Otherwise, the user can specify
that they should be drawn with replacement if the number of imputed datasets
is to be greater than the number of chains.

In the case of non-replacement; this function is a little less naive than
applying `sample.int` to each row - instead if the number of observations is
large enough it first generates the possible arrangements for each row, and
then draws from the arrangements (see  `utils::combn`).

When sampling with replacement, the implementation is a trivial call to
`sample.int`.

Completed data sets as long data.frame

Completed (imputed) data sets given incomplete data passed to
\code{\link{smirf}} and the output thereof, returned in a long-format
data.frame.

Description
}
\section{To-do}{

\itemize{
  \item document \code{complete_results}
  \item write tests for \code{complete_results}
}
}

\seealso{
\code{\link{smirf}}
}
