% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{frequency_of_imputed}
\alias{frequency_of_imputed}
\title{Frequency of imputed values of categorical data}
\usage{
frequency_of_imputed(imputed, to_categories = list())
}
\arguments{
\item{imputed}{named list;}

\item{to_categories}{named list;}
}
\description{
Frequency of imputed values of categorical data
}
\section{To-do}{

\itemize{
  \item fully document \code{frequency_of_imputed}
  \item write tests for \code{frequency_of_imputed}
}
}

\keyword{internal}
