% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_naive.R
\name{impute_naive}
\alias{impute_naive}
\title{Impute missing data naively using complete cases}
\usage{
impute_naive(
  data,
  indicator = lapply(data, is.na),
  mode = c("sample", "aggregate")
)
}
\arguments{
\item{indicator}{list (named); each item is the indicator (vector) of missing
values for a column in the data.}
}
\value{
data.frame; an imputed data set with the same structure as the \code{data}
argument.
}
\description{
The initial step in multiple imputation with random forests requires missing
values to be drawn \emph{without} using the model; in the case of multiple
imputation via chained equations the complete cases are sampled from with
replacement (van Buuren and Groothuis-Oudshoorn, 2012), whereas for the
missForest algorithm, the most-frequent value or mean is used (Stekhoven and
Buehlmann, 2012).
}
\details{
Two wrappers are provided for this function so that it can be passed as
\code{fn_init} to \code{\link[=smirf]{smirf()}}, these are \code{\link[=impute_naive_by_sample]{impute_naive_by_sample()}} and
\code{\link[=impute_naive_by_aggregate]{impute_naive_by_aggregate()}}. In either case, the function will modify each
column (in order) by replacing missing values with either a sample from the
complete cases, or an aggregate of the complete cases (such as the most
frequent value).

Sampling generates data with higher entropy than given by the aggregate
approach. When using a stopping condition, care must be taken that the choice
of initial state doesn't introduce a false positive 'stop' in early
iterations. We suspect that the 'aggregate' approach will results in better
specificity of the stop condition.
}
\examples{
impute_naive(data.frame(x=c(0,1,NA)), mode='sample')
impute_naive(data.frame(x=c(0,1,NA)), mode='aggregate')

}
\references{
\itemize{
\item Stekhoven, D.J. & Buehlmann, P. (2012). MissForest--non-parametric
missing value imputation for mixed-type data. \emph{Bioinformatics, 28}(1),
112-118. \url{doi:10.1093/bioinformatics/btr597}.
\item Van Buuren, S. & Groothuis-Oudshoorn, K. (2011). mice: Multivariate
Imputation by Chained Equations in R. \emph{Journal of Statistical Software,
45}(3), 1-67. \url{doi:10.18637/jss.v045.i03}.
}
}
\seealso{
\code{\link[=smirf]{smirf()}} \code{\link[=impute_naive_by_sample]{impute_naive_by_sample()}} \code{\link[=impute_naive_by_aggregate]{impute_naive_by_aggregate()}}
\code{\link[mice:mice]{mice::mice()}} \code{\link[missForest:missForest]{missForest::missForest()}}
}
