% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_naive.R
\name{impute_naive_by_sample}
\alias{impute_naive_by_sample}
\title{Impute missing data naively by sampling from complete cases.}
\usage{
impute_naive_by_sample(data, indicator = lapply(data, is.na))
}
\arguments{
\item{data}{data.frame: an incomplete data set of numeric, logical, integer,
factor, or ordered data - one row per observation, one column per variable.}

\item{indicator}{list: an indicator of missing status for each (named)
variable for each row in the data, equal to 1 (or \code{TRUE}) for a missing case,
otherwise 0 (or \code{FALSE}).}
}
\value{
data.frame: an imputed data set with the same structure as the \code{data}
argument.
}
\description{
Drawing imputed values without using the fully conditional specification (or
other joint model) is performed during the first step of the 'mice'
imputation algorithm (van Buuren and Groothuis-Oudshoorn, 2012). In this
case sampling from complete cases is used.
}
\details{
This function can be passed directly to \code{\link[=mimputest]{mimputest()}} as the value of the
\code{fn_init} argument. This function will iterate over each column in the data
and replace missing values by sampling (with replacement) from the complete
cases.

This approach may improve the sensitivity (but lower the specificity) in some
stopping conditions used by the main algorithm loop during early iterations,
owing to the entropy being closer to that of the converged process.
}
\examples{
\dontrun{
mimputest(iris, fn_init=impute_naive_by_aggregate)
}
impute_naive_by_sample(data.frame(x=c(0,1,NA)))

}
\references{
\itemize{
\item Van Buuren, S. & Groothuis-Oudshoorn, K. (2011). mice: Multivariate
Imputation by Chained Equations in R. \emph{Journal of Statistical Software},
\emph{45}(3), 1-67. \doi{10.18637/jss.v045.i03}.
}
}
\seealso{
\code{\link[=impute_naive]{impute_naive()}}
}
