% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leiks_D.R
\name{leiks_D}
\alias{leiks_D}
\title{Calculate Leik's D of ordered data}
\usage{
leiks_D(x, freq = table(as.ordered(x)), cdf = cumsum(freq)/sum(freq))
}
\arguments{
\item{x}{ordered;
observed ordered data or data that can be coerced into ordered
via as.ordered.}

\item{freq}{numeric;
count of observations in each category in order.}

\item{cdf}{numeric;
cumulative proportion of observations in each category in order.}
}
\value{
numeric;
            scalar value of Leik's D
}
\description{
Calculate Leiks' D of ordered data using tally of observations in each
category (provided in order)
}
\section{To-do}{

\itemize{
  \item document \code{entropy}.
  \item write tests for \code{entropy}.
}
}

\keyword{internal}
