% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure_correlation}
\alias{measure_correlation}
\title{Measure the association between sequential complete data sets.}
\usage{
measure_correlation(x_sample, y_sample, data, indicator, method = "kendall")
}
\arguments{
\item{x_sample}{named list: the imputed values of the missing cases within
each variable from one iteration of the missForest algorithm.}

\item{y_sample}{named list: the imputed values from the following (or
preceding) iteration for \code{x_sample} with the same format as \code{x_sample}.}

\item{data}{data.frame: an incomplete data set of numeric, logical, integer,
factor, or ordered data - one row per observation, one column per variable.}

\item{indicator}{list: an indicator of missing status for each (named)
variable for each row in the data, equal to 1 (or \code{TRUE}) for a missing case,
otherwise 0 (or \code{FALSE}).}

\item{method}{character: the correlation coefficient to calculate for
ordered data passed to \code{\link[stats:cor]{stats::cor()}}.}
}
\value{
named numeric: \code{measure} equal to the mean over all variables of the
(rank) correlation or Cramer's V for ordered and non-ordered data
(respectively).
}
\description{
The association (or correlation) of the data from one complete data set to
another can be used to assess stopping criterion for the missForest procedure
(Stekhoven and Buehlmann, 2012). Ordered data associations are measured by
rank correlations; while non-ordered data associated are measures by
Cramer's V with bias correction
\url{https://en.wikipedia.org/wiki/Cram\%C3\%A9r\%27s_V}. The mean across all
variables is returned as a single measure.
}
\details{
The missForest algorithm compares two sequential measures, i.e. the
association between the most recent and second-most recent data sets versus
the association between the second and third-most recent data sets. When the
measures have changed direction, the algorithm stops. It is assumed that the
measures are initially increasing, and when the measure decreases the
algorithm has converged. When using association or correlation measures, this
is approximately when the entropy has reached a maximum.

The type of (rank) correlation can be changed via the \code{measure} argument,
see \code{\link[stats:cor]{stats::cor()}} for details of the correlation measures available. The
Pearson measure is not recommended for ordered data but this function will
not error or warn if you select it - the ordered variable will be transformed
via \code{\link[=xtfrm]{xtfrm()}} and treated as continuous for the purpose of the Pearson
correlation calculation.
}
\examples{
\dontrun{
# simply pass to mimputest
mimputest(iris, stop_measure=measure_correlation)
}
}
\references{
\itemize{
\item Stekhoven, D. J. & Buehlmann, P. (2012). MissForest--non-parametric
missing value imputation for mixed-type data. \emph{Bioinformatics}, 28(1),
112-118. \doi{10.1093/bioinformatics/btr597}.
}
}
\seealso{
\code{\link[=sampler_loop]{sampler_loop()}} \code{\link[stats:cor]{stats::cor()}}
}
