% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smirf.R
\name{smirf}
\alias{smirf}
\title{Single of multiple imputation of missing data using random forests}
\usage{
smirf(X, n = 5L, order.impute = NULL, gibbs = F, tree.imp = F,
  boot.train = F, obs.only = T, verbose = F,
  X.init.fn = no_information_impute,
  stop.measure = measure_correlation, loop.limit = 10L,
  overrides = list(), clean.step = list(), ...)
}
\arguments{
\item{X}{data.frame;
a incomplete data set including any of numeric, logical, integer,
factor and ordered data types.}

\item{n}{numeric scalar;
the number of imputations - i.e. number of times the missForest
algorithm is used.}

\item{order.impute}{character vector;
(optional) order to impute columns - default value given by
increasing rate of missing data in \code{X}.}

\item{gibbs}{logical;
use Gibbs sampling in training steps (\code{T}) rather than the
predictions from the previous iteration (default).}

\item{tree.imp}{logical;
use a prediction of missing data from single tree in the forest 
when training (\code{T}) rather than the bagged predicted value
(default).}

\item{boot.train}{logical;
train each forest on a bootstrap sample of the observed data
when \code{T}, rather than the observed data (default).}

\item{obs.only}{logical;
train on only observed outcomes (default) or use all data
including predicted/sampled values of missing outcomes (\code{T}).}

\item{verbose}{logical;
print additional output.}

\item{X.init.fn}{function;
creates a completed data set to be used as the initial state of
the missForest procedure given two arguments; \itemize{
    \item a data.frame
    \item a list with and item indicating the missing (\code{T})
          or not-missing (\code{F}) status of each column of the
          first argument,
} the default \code{\link{no_information_impute}} serves as an
example.}

\item{stop.measure}{function;
evaluates the difference or relationship between the two most
recently completed data sets during iteration, must accept the
following arguments;
\describe{
     \item{\code{X}}{named list with imputed values (in order of
         appearance by row) for each column in the data set;}
     \item{\code{Y}}{named list with imputed values (in order of
         appearance by row) for each column in the data set;}
     \item{\code{X_init}}{the original (mised-type) data set with
          missing values replaced as at the starting point of
          missForest;}
     \item{\code{indicator}}{a list with the missing (\code{=T})
          or not missing (\code{=F}) status of the original data
          set;}
} and should return a numeric (vector), the default
\code{\link{measure_correlation}} serves as an example, or see the
original measure proposed by Stekhoven and Buehlmann (2012) in
\code{\link{measure_stekhoven_2012}}.}

\item{loop.limit}{numeric;
maximum number of iterations within missForest procedure.}

\item{overrides}{named list;
(variable-wise) over-rides for arguments passed to
\code{\link[ranger]{ranger}} when training on the response
variable given by the name of the item.}

\item{clean.step}{named list;
each item is a function to clean or post-process the named imputed
data immediately after it is imputed, taking two arguments;
\itemize{
    \item the subset of the data used in the current training step
          which had missing values of the named data,
    \item the most recently imputed values of the named data,
} and should return (post-processed) data of the same length and
type as the second argument.}

\item{...}{further arguments passed to all calls to 
\code{\link[ranger]{ranger}}, e.g. \code{num.trees} for the number
of trees in each forest.}
}
\description{
Missing data (multiple) imputation using the missForest algorithm by
Stekhoven and Buehlmann (2012). The ranger (Wright and Ziegler, 2017) fast
implementation of random forest (training) algorithm is used. Some
key alterations to the missForest algorithm may be specified by the user.
}
\details{
For a full description of the missForest algorithm, see Stekhoven and
Buehlmann (2012). In brief, at each iteration missing values are imputed for
each variable (in the order of \code{order.impute}) by the predictions of a
random forest trained on the observed cases of that variable along with the
completed data set of the previous iteration as the value of the predictors.
This is repeated until some measure of the relationship between iterations
indicates convergence - usually by decreasing from the measure at the
previous iteration.

By default the columns are imputed in the order of least missing to most
missing. This can be over-ridden by the \code{order.impute} argument. Columns
that are entirely missing are excluded. Non-integer numeric data is treated
as continuous and predicted by regression forests while all other data,
including integer and logical data, are predicted via classification forests.
No special treatment is given to ordered categorical data.

The call to \code{\link[ranger]{ranger}} may be modified by the \code{...}
arguments, and any variable-specific argument to pass may be specified in the
\code{overrides} argument.

The key modifications to the missForest procedure governed by the arguments
\describe{
    \item{\code{gibbs}}{use the most recent predictions for each variable
        in training and prediction as they become available, like a Gibbs
        sampler by setting this to \code{T} (default is \code{F};}
    \item{\code{tree.imp}}{predict using a randomly selected tree for each
        missing value rather than use the whole-of-forest aggregated
        prediction by setting this to \code{T} (default is \code{F}),}
    \item{\code{boot.train}}{train on a boot-strapped resample of the data,
        default is \code{F}), and;}
    \item{\code{obs.only}}{train on all rows in the data set instead of
        observed only by setting this to \code{F} (default is \code{T}).}
}

Using the first two changes make the procedure similar to the Multiple
Imputation via Chained Equations of Doove et al (2014). Experimentally,
adding the final modification makes the procedure mimic van Buuren and
Groothuis-Oudshoorn, (2012), however it does not seem advisable to train a
random forest on predicted data. The third item follows the suggestion of
Bartlett (2014) to improve CI coverage.

The convergence criterion can be modified by the \code{stop.measure}
argument. The default is to measure the mean rank correlation between
iterations of the ordered data and the stationary rate of the categorical
data (see \code{\link{measure_correlation}}. The procedure halts when both of
these values are less than or equal to the previous values (see
\code{\link{stop_condition}}). The original Stekhoven and Buehlmann (2012)
measure is provided by the \code{\link{measure_stekhoven_2012}} function.
}
\section{To-do}{


The post-processing of the output needs to be determined. It may be useful to
prepare a \code{\link[mice]{mids}} object so that the powerful tools that
have been developed there might be used on data sets generated by smirf.

\itemize{
  \item consider run-time like tests for \code{stop.measure}
  \item document \code{smirf} further
}

Need to consider a variable-by-variable approach to specifying imputation
model.
}

\references{
Bartlett, J., 2014. 'Methodology for multiple imputation for missing data in
electronic health record data', presented to _27th International Biometric
Conference_, Florence, July 6-11.

Doove, L.L., Van Buuren, S. and Dusseldorp, E., 2014. Recursive partitioning
for missing data imputation in the presence of interaction effects. 
\emph{Computational Statistics & Data Analysis, 72}, pp. 92-104.
\href{https://dx.doi.org/10.1016/j.csda.2013.10.025}{doi.10.1016/j.csda.2013.10.025}

Stekhoven, D.J. and Buehlmann, P., 2012. MissForest--non-parametric
missing value imputation for mixed-type data. \emph{Bioinformatics, 28}(1),
pp. 112-118.
\href{https://dx.doi.org/10.1093/bioinformatics/btr597}{doi.1.1093/bioinformatics/btr597}

Wright, M. N. and Ziegler, A., 2017. ranger: A fast implementation of random
forests for high dimensional data in C++ and R. \emph{Journal of Statistical
Software, 77}(i01), pp. 1-17. \href{https://dx.doi.org/10.18637/jss.v077.i01}{doi.10.18637/jss.v077.i01}
}
\seealso{
\code{\link{measure_correlation}}
         \code{\link{measure_stekhoven_2012}}
         \code{\link{stop_condition}}
         \code{\link{no_information_impute}} \code{\link{sample_impute}}
         \code{\link[missForest]{missForest}}
         \code{\link[ranger]{ranger}}
}
