% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_of_imputed.R
\name{statistics_of_imputed}
\alias{statistics_of_imputed}
\title{Calculate statistics of imputed data}
\usage{
statistics_of_imputed(X_predict)
}
\arguments{
\item{X_predict}{named list;}
}
\description{
Calculates certain selected statistics (aggregates) of imputed data
}
\section{To-do}{

\itemize{
  \item fully document \code{statistics_of_imputed}
  \item write tests for \code{statistics_of_imputed}
}
}

\keyword{internal}
